Consolidate integral types. (#67)
* Use int for index type.

* Address warnings.

* .clang-tidy format was wrong.

* Done except assert(...).

* Remove debug messages.

* Done.

* Clean up.

+++ include/polatory/rbf/biharmonic2d.hpp
@@ -57,7 +57,7 @@ public:
     }
   }
 
-  size_t num_parameters() const override {
+  int num_parameters() const override {
     return 1;
   }
