+++ include/polatory/rbf/biharmonic3d.hpp
@@ -48,7 +48,7 @@ public:
     *gradz = c * z;
   }
 
-  size_t num_parameters() const override {
+  int num_parameters() const override {
     return 1;
   }
