+++ include/polatory/rbf/reference/triharmonic3d.hpp
@@ -49,7 +49,7 @@ public:
     *gradz = c * z;
   }
 
-  size_t num_parameters() const override {
+  int num_parameters() const override {
     return 1;
   }
